django_rest_frame_api/apis/serializers.py
from rest_framework import serializers
from .models import City, Users


class CitySerializer(serializers.ModelSerializer):
    class Meta:
        model = City
        fields = '__all__'


class UsersSerializer(serializers.ModelSerializer):
    class Meta:
        model = Users
        fields = '__all__'
